[refactor] 파일 위치 변경에 따른 참조 수정 및 스포츠 배치 로직 최적화
